server_flask/requirements.txt
certifi
chardet
click
cmake
decorator
dlib
face-recognition
face-recognition-models
Flask
Flask-Cors
idna
itsdangerous
Jinja2
MarkupSafe
imageio
imageio-ffmpeg
moviepy
matplotlib
numpy
opencv-python
Pillow
proglog
psycopg2
python-dotenv
requests
six
tqdm
urllib3
Werkzeug
wincertstore